be/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.0
google-auth==2.25.2
python-dotenv==1.0.0
gunicorn==21.2.0
Flask-Limiter==3.5.0
